requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.5
django-import-export==2.5.0
et-xmlfile==1.1.0
filelock==3.0.12
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
platformdirs==2.0.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
virtualenv==20.6.0
waitress==2.0.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
